+++ CHANGELOG.md
@@ -1,13 +1,21 @@
 # PickupArtist
 
+## Unreleased
+
+- Update mod to support game version 1.20.7
+  - Fix piles no longer stacking higher than one block
+
 ## 0.2.0
+
 - Change how source and destination slots are determined
 - Change event handling to improve compatibility
 - Add animation and sound when taking from crates
 - Resolve build issues caused by build optimizations
 
 ## 0.1.1
+
 - Reupload due to build and packaging issues
 
 ## 0.1.0
+
 - Initial release